From Debian: Accept 0 as a threshold from the command-line
Bug text:

Package: joy2key
Version: 1.6.1
Tags: patch

I own a gamepad, for which the axis belonging to the steering cross only
report values -1,0,1 so the threshold has to be set to 0, which works
when using the build-in calibrate function. The problem is, that when
trying to specify this via the commandline, this is not accepted,
because the values are tested as truth value and 0 evaluates to
false ...

To fix this I created an additional array, holding just the information,
whether the threshold for this axis was defined. To allow manual
calibration, I introduced "x" as placeholder in the threshold
commandline, meaning: "This threshold is not supplied", which was prior
to this work indicated by passing 0 as value.

Please see the attached patch

Thanks

Matthias